requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
pandas           
pydantic
python-dotenv